requirements.txt
absl-py==2.3.1
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
asttokens==3.0.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
captum==0.8.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fastapi==0.128.0
fastjsonschema==2.21.2
filelock==3.20.1
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
gitdb==4.0.12
GitPython==3.1.46
grpcio==1.76.0
h11==0.16.0
holidays==0.87
idna==3.11
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.3
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jupyter_client==8.7.0
jupyter_core==5.9.1
kiwisolver==1.4.9
lightning-utilities==0.15.2
Markdown==3.10
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
missingno==0.5.2
mpmath==1.3.0
multidict==6.7.0
narwhals==2.14.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.6.1
neuralprophet==0.8.0
numpy==1.26.4
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
packaging==25.0
pandas==2.3.3
pandas-stubs==2.3.3.251219
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
plotly==5.24.1
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==6.33.2
psutil==7.2.1
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==22.0.0
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
pyparsing==3.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytorch-lightning==1.9.5
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
scipy==1.16.3
seaborn==0.13.2
setuptools==80.9.0
six==1.17.0
smmap==5.0.2
stack-data==0.6.3
starlette==0.50.0
streamlit==1.52.2
sympy==1.14.0
tenacity==9.1.2
tensorboard==2.20.0
tensorboard-data-server==0.7.2
toml==0.10.2
torch==2.9.1
torchmetrics==1.8.2
tornado==6.5.4
tqdm==4.67.1
traitlets==5.14.3
triton==3.5.1
types-pytz==2025.2.0.20251108
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.2
uvicorn==0.40.0
watchdog==6.0.0
wcwidth==0.2.14
Werkzeug==3.1.4
yarl==1.22.0
